full_demo/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "augmentation_agent"
version = "0.1.0"
description = "Lab 7: Graph Augmentation Agent — DSPy structured analysis for Neo4j enrichment"
requires-python = ">=3.12"
dependencies = [
    "dspy>=3.0.4",
    "databricks-sdk",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "neo4j>=5.0",
]

[dependency-groups]
dev = [
    "databricks-job-runner",
]


[tool.setuptools.packages.find]
include = ["augmentation_agent*"]
